% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_sierraleone_2014.R
\docType{data}
\name{ebola_sierraleone_2014}
\alias{ebola_sierraleone_2014}
\title{Ebola virus disease outbreak in Sierra Leone, 2014--2015}
\format{
A data frame with columns:
\describe{
\item{\code{id}}{Numeric case identifier.}
\item{\code{age}}{Age of the case (years).}
\item{\code{sex}}{Sex of the case (\code{"F"} or \code{"M"}).}
\item{\code{status}}{Case classification: \code{"confirmed"} or
\code{"suspected"}.}
\item{\code{date_of_onset}}{Date of symptom onset}
\item{\code{date_of_sample}}{Date the diagnostic sample was collected.}
\item{\code{district}}{District of the case.}
\item{\code{chiefdom}}{Chiefdom.}
}

A data frame.
}
\source{
Data from the  \code{outbreaks} R package.
\url{https://github.com/reconhub/outbreaks}
}
\usage{
data(ebola_sierraleone_2014)
}
\description{
Line-list of confirmed and suspected Ebola virus disease cases from
Sierra Leone from 2014-05-18 to 2015-09-13. Data from the
\code{outbreaks} package.
}
\examples{
data(ebola_sierraleone_2014)
}
\references{
L. Fang, et al. 2016. \emph{Ebola virus disease in Sierra Leone.} Proceedings of
the National Academy of Sciences, 113 (16) 4488-4493;
DOI: 10.1073/pnas.1518587113
}
\seealso{
Other real-data: 
\code{\link{denguedat}},
\code{\link{flu_h7n9_china_2013}},
\code{\link{flusight}},
\code{\link{germany_covid19_hosp}},
\code{\link{measles_hagelloch_1861}},
\code{\link{mers_korea_2015}},
\code{\link{mpoxdat}}
}
\author{
Data from Fang et al. (2016) Transfer to R and documentation by Finlay
Campbell (finlaycampbell93@gmail.com).
}
\concept{real-data}
\keyword{ebola}
