% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denguedat.R
\docType{data}
\name{denguedat}
\alias{denguedat}
\title{denguedat: Dengue fever individual-level reporting data from Puerto Rico}
\format{
A data frame.
}
\usage{
data(denguedat)
}
\description{
Surveillance data from CDC Division of Vector-Borne Diseases.
1990-2010 case reporting data included.
}
\details{
Each row represents a case with the columns indicating the following:
\itemize{
\item \code{onset_week}:  the week of symptom onset.
\item \code{report_week}: the week of case report.
}
}
\note{
Data originally from the \code{NobBS} package. While \code{onset_week} and \code{report_week}.
The package contained a synthetic gender column which was removed here
to keep just the real data.
}
\examples{
data(denguedat)
}
\references{
MCGOUGH, Sarah F., et al. Nowcasting by Bayesian Smoothing: A flexible, generalizable model for
real-time epidemic tracking. PLoS computational biology, 2020, vol. 16, no 4, p. e1007735.
}
\seealso{
Other real-data: 
\code{\link{flusight}},
\code{\link{germany_covid19_hosp}},
\code{\link{mpoxdat}}
}
\concept{real-data}
\keyword{dengue}
