% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flusight.R
\docType{data}
\name{flusight}
\alias{flusight}
\title{USA's National Healthcare Safety Network (NHSN) Weekly Hospital Respiratory Data from FluSight}
\format{
A data frame.
}
\source{
\url{https://github.com/cdcepi/FluSight-forecast-hub/blob/main/target-data/time-series.csv}
}
\usage{
data(flusight)
}
\description{
FluSight's weekly hospital admission prediction targets
based on the 'total number of new hospital admissions of patients with confirmed
influenza captured during the reporting week' reported through CDC's NHSN
(the dataset formerly known as HHS-Protect), Weekly Hospital Respiratory Data.
Data was downloaded on November 12th 2025.
}
\details{
Data represents how many cases were considered influenza during the week of
\emph{target_end_date} given the information known until week \emph{as_of}. Note
that \emph{as_of} is always one week ahead of \emph{target_end_date}.

This is count data. Each row represents the reported case count for a given location, target
week, and reporting snapshot. The columns are as follows:
\itemize{
\item \code{as_of}: The date on which the data snapshot was taken (one week after \code{target_end_date}),
formatted in ISO8601 as YYYY-MM-DD.
\item \code{target_end_date}: The Saturday ending the surveillance week, formatted in ISO8601 as YYYY-MM-DD.
\item \code{observation}: Case counts for that target week as of \code{as_of}.
\item \code{location_name}: Name of the US state, district, or territory.
}
}
\examples{
data(flusight)
}
\references{
Target data from Flusight. Online:
\url{https://github.com/cdcepi/FluSight-forecast-hub/blob/main/target-data/time-series.csv}
}
\seealso{
Other real-data: 
\code{\link{denguedat}},
\code{\link{ebola_sierraleone_2014}},
\code{\link{flu_h7n9_china_2013}},
\code{\link{germany_covid19_hosp}},
\code{\link{measles_hagelloch_1861}},
\code{\link{mers_korea_2015}},
\code{\link{mpoxdat}}
}
\concept{real-data}
\keyword{flu}
\keyword{influenza}
