% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germany_covid19_hosp.R
\docType{data}
\name{germany_covid19_hosp}
\alias{germany_covid19_hosp}
\title{germany_covid19_hosp: Incident COVID-19 hospitalisations indexed by the date of positive test
(reference date) and report date from Germany in 2021 and 2022.}
\format{
A \code{data.frame} with 140,630 rows and 6 columns.
\describe{
\item{\code{reference_date}}{Date of first positive COVID-19 test formatted
in ISO8601 standards as YYYY-MM-DD.}
\item{\code{location}}{Character string indicating the location of the case
counts}
\item{\code{age_group}}{Character string indicating the age group of the case
counts.}
\item{\code{delay}}{Integer specifying the delay, in days, between the reference
date and the report date}
\item{\code{count}}{Integer indicating the number of cases indexed by reference
and report date.}
\item{\code{report_date}}{Date of case report, formatted in ISO8601 standards as
YYYY-MM-DD.}
}

A data frame.
}
\source{
This data comes directly from the preprocessed data in the
German COVID-19 Nowcast Hub from \url{https://github.com/KITmetricslab/hospitalization-nowcast-hub/blob/main/data-truth/COVID-19/COVID-19_hospitalizations_preprocessed.csv}
It contains incident case counts by age group in Germany.
}
\usage{
germany_covid19_hosp
}
\description{
germany_covid19_hosp: Incident COVID-19 hospitalisations indexed by the date of positive test
(reference date) and report date from Germany in 2021 and 2022.
}
\references{
Johnson KE, Tyszka E, Bracher J, Funk S, Abbott S (2025).
\emph{baselinenowcast: Methods for baseline nowcasting
right-truncated epidemiological data}. \url{https://github.com/epinowcast/baselinenowcast/}.
}
\seealso{
Other real-data: 
\code{\link{denguedat}},
\code{\link{flusight}},
\code{\link{mpoxdat}}
}
\concept{real-data}
\keyword{covid}
