% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcastdatasets.R
\docType{package}
\name{nowcastdatasets-package}
\alias{nowcastdatasets}
\alias{nowcastdatasets-package}
\title{nowcastdatasets: Datasets for Epidemiological Nowcasting Models}
\description{
Includes different datasets for epidemiological nowcasting.
}
\author{
\strong{Maintainer}: Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com} (\href{https://orcid.org/0000-0003-4471-5270}{ORCID})

Other contributors:
\itemize{
  \item Sarah McGough \email{sfm341@mail.harvard.edu} (dengue dataset adapted from NobBS) [contributor]
  \item Nicolas Menzies \email{nmenzies@hsph.harvard.edu} (dengue dataset adapted from NobBS) [contributor]
  \item Marc Lipsitch \email{mlipsitc@hsph.harvard.edu} (dengue dataset adapted from NobBS) [contributor]
  \item Michael Johansson \email{mjohansson@cdc.gov} (dengue dataset adapted from NobBS) [contributor]
  \item Rami Yaari \email{ramiyaari@gmail.com} (dengue dataset adapted from NobBS) [contributor]
  \item Teresa Yamana \email{tky2104@climate.columbia.edu} (dengue dataset adapted from NobBS) [contributor]
  \item Matteo Perini \email{mperini@climate.columbia.edu} (dengue dataset adapted from NobBS) [contributor]
  \item Rafael Lopes \email{rafael.lopes@yale.edu} (\href{https://orcid.org/0000-0002-9416-6145}{ORCID}) (SARI_BH dataset adapted from nowcaster.) [contributor]
  \item Tatiana Portella \email{tatiana.portella@fiocruz.br} (\href{https://orcid.org/0000-0001-8364-1978}{ORCID}) (SARI_BH dataset adapted from nowcaster.) [contributor]
  \item Marcelo Gomes \email{marcelo.gomes@fiocruz.br} (\href{https://orcid.org/0000-0003-4693-5402}{ORCID}) (SARI_BH dataset adapted from nowcaster.) [contributor]
  \item Leonardo Bastos \email{leonardo.bastos@fiocruz.br} (\href{https://orcid.org/0000-0002-1406-0122}{ORCID}) (SARI_BH dataset adapted from nowcaster.) [contributor]
  \item Kaitlyn Johnson \email{kaitlyn.johnson@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8011-0012}{ORCID}) (Germany hospitalizations data from baselinenowcast) [contributor]
  \item Emily Tyszka \email{etsyzka@umass.edu} (\href{https://orcid.org/0009-0005-6088-4017}{ORCID}) (Germany hospitalizations data from baselinenowcast) [contributor]
  \item Johannes Bracher \email{johannes.bracher@kit.edu} (\href{https://orcid.org/0000-0002-3777-1410}{ORCID}) (Germany hospitalizations data from baselinenowcast) [contributor]
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) (Germany hospitalizations data from baselinenowcast) [contributor]
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID}) (Germany hospitalizations data from baselinenowcast) [contributor]
  \item Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) (Germany hospitalizations data from baselinenowcast) [contributor]
}

}
\keyword{internal}
